Modularize cache modules for gate compliance
